+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## New functionalities
 
+* `subset` and `subset!` now allow passing zero column selectors
+   ([#3025](https://github.com/JuliaData/DataFrames.jl/pull/3025))
 * `permutedims` now supports a `strict` keyword argument that allows
   for a more flexible handling of values stored in a column that will
   become a new header
@@ -9,6 +11,8 @@
 * `unstack` now allows passing a function in `valuestransform` keyword argument;
   this allows for a convenient creation of two dimensional pivot tables
   ([#2998](https://github.com/JuliaData/DataFrames.jl/issues/2998))
+* `filter` for `GroupedDataFrame` now accepts `ungroup` keyword argument
+  ([#3021](https://github.com/JuliaData/DataFrames.jl/issues/3021))
 * Add special syntax for `eachindex`, `groupindices`, and `proprow`
   to transformation mini-language
   ([#3001](https://github.com/JuliaData/DataFrames.jl/pull/3001)).
@@ -17,6 +21,13 @@
   ([#3010](https://github.com/JuliaData/DataFrames.jl/pull/3010)).
 * `first` and `last` for `GroupedDataFrame` now support passing number of elements to get
   ([#3006](https://github.com/JuliaData/DataFrames.jl/issues/3006))
+* Add `insertcols`, which is a version of `insertcols!` that creates a new data frame
+  ([#3020](https://github.com/JuliaData/DataFrames.jl/issues/3020))
+* Add `fillcombinations` function that generates all combinations of
+  levels of selected columns of a data frame
+  ([#3012](https://github.com/JuliaData/DataFrames.jl/issues/3012))
+* Guarantee that `permute!` and `invpermute!` throw on invalid input
+  ([#3035](https://github.com/JuliaData/DataFrames.jl/pull/3035))
 * New experimental functions `DataFrames.singlethreaded` and `DataFrames.setmultithreading`
   allow disabling multithreading in all DataFrames operations
   ([#3030](https://github.com/JuliaData/DataFrames.jl/pull/3030))
@@ -27,6 +38,12 @@
   into an existing column of a data frame replaces it. Under Julia 1.6
   or older it is an in place operation.
   ([#3022](https://github.com/JuliaData/DataFrames.jl/pull/3022)
+
+## Performance
+
+* Speed up `permute!` and `invpermute!` (and therefore sorting) 2x-8x 
+  for large tables by using cycle notation
+  ([#3035](https://github.com/JuliaData/DataFrames.jl/pull/3035))
 
 # DataFrames.jl v1.3.2 Patch Release Notes